fastapi-backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgi-lifespan==2.1.0
certifi==2025.11.12
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.124.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
nodeenv==1.9.1
pdfminer.six==20231228
prisma==0.15.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.20
starlette==0.49.3
StrEnum==0.4.15
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
pytest==8.0.0
pytest-asyncio==0.23.5
reportlab==4.1.0
